requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
neo4j
redis
langgraph
langchain
langchain-core
openai
tavily-python
httpx
python-dotenv
python-multipart
tenacity
structlog
orjson
pgvector
jinja2
pytest
pytest-asyncio
pandas
numpy
scikit-learn
python-jose[cryptography]
passlib
email-validator
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2